Add equal() method
This adds an equal() method to Absolute, Date, DateTime, MonthDay, Time,
and YearMonth. Currently `a.equal(b)` is equivalent to
`Temporal.X.compare(a, b) === 0`, but after calendars are added, equal()
should additionally return false if the calendars in the two objects are
different.

Duration doesn't have an equal() method at this time because, does
`{ minutes: 1, seconds: 30 }` equal `{ seconds: 90 }`? TimeZone doesn't
have one either because it's simple enough to compare the .name
property.